Fixed clear data button.... Oopsies
