+++ composeApp/src/commonMain/kotlin/nodes/TeleNode.kt
@@ -14,7 +14,6 @@ class TeleNode (
     private val backStack: BackStack<AutoTeleSelectorNode.NavTarget>,
 
     private val selectAuto: MutableState<Boolean>,
-    private val lostComms: MutableState<Boolean>,
 
     private val match: MutableState<String>,
     private val team: MutableIntState,
@@ -22,6 +21,6 @@ class TeleNode (
 ) : Node(buildContext) {
     @Composable
     override fun View(modifier: Modifier) {
-        TeleMenu(backStack, selectAuto, lostComms, match, team, robotStartPosition)
+        TeleMenu(backStack, selectAuto, match, team, robotStartPosition)
     }
 }
