+++ composeApp/src/desktopMain/kotlin/FileMaker.desktop.kt
@@ -42,3 +42,8 @@ fun exportScoutData() {
     writer.write(gson.toJson(matchScoutArray))
     writer.close()
 }
+fun deleteFile(){
+    val homeDir = System.getProperty("user.home")
+    val file = File("$homeDir/Documents/match_scouting_data.json")
+    file.delete()
+}
